+++ src/commands/buildSoy.ts
@@ -5,7 +5,10 @@ import { logger } from "../logger";
 import { compileSoy } from "../soy";
 
 export type BuildSoyConfig = Required<
-  Pick<DuckConfig, "soyFileRoots" | "soyJarPath" | "soyOptions">
+  Pick<
+    DuckConfig,
+    "soyFileRoots" | "soyJarPath" | "soyClasspaths" | "soyOptions"
+  >
 >;
 
 /**
@@ -24,9 +27,9 @@ export async function buildSoy(
 }
 
 async function findSoyFiles(config: BuildSoyConfig): Promise<string[]> {
-  const soyFilePromises = config.soyFileRoots.map(async p => {
+  const soyFilePromises = config.soyFileRoots.map(async (p) => {
     const files = await recursive(p);
-    return files.filter(file => /\.soy$/.test(file));
+    return files.filter((file) => /\.soy$/.test(file));
   });
   const soyFiles = flat(await Promise.all(soyFilePromises));
   return soyFiles;