+++ src/reporters/text-reporter.ts
@@ -11,7 +11,7 @@ export class TextReporter extends BaseReporter {
     super({
       stderr: options.stderr || true,
       outputDir: options.outputDir || null,
-      resultFilename: "results.txt"
+      resultFilename: "results.txt",
     });
   }
 
@@ -19,13 +19,13 @@ export class TextReporter extends BaseReporter {
     // if items doesn't include any errors or warnings,
     // we ignore "info" that includes the line of summary like:
     // "0 error(s), 0 warning(s), xx% typed"
-    if (items.filter(i => i.level !== "info").length === 0) {
+    if (items.filter((i) => i.level !== "info").length === 0) {
       return "";
     }
     return `# ${entryConfigPath}:
 ${command ? `\n${command}\n` : ""}
 ${items
-  .map(item =>
+  .map((item) =>
     item.level === "info" ? item.description : this.formatErrorCase(item)
   )
   .join("\n\n")}\n`;