+++ test/text-reporter.ts
@@ -14,7 +14,7 @@ describe("TextReporter", () => {
       const actual = reporter.format({
         entryConfigPath,
         command,
-        items: []
+        items: [],
       });
       assert.equal(actual, "");
     });
@@ -26,9 +26,9 @@ describe("TextReporter", () => {
         items: [
           {
             level: "info",
-            description: "0 error(s), 0 warning(s), 98.4% typed"
-          }
-        ]
+            description: "0 error(s), 0 warning(s), 98.4% typed",
+          },
+        ],
       });
       assert.equal(actual, "");
     });
@@ -48,9 +48,9 @@ describe("TextReporter", () => {
             line: 57,
             column: 32,
             context:
-              "  this.outstandingEvents_ = new goog.structs.Map();\n                                ^^^^^^^^^^^^^^^^"
-          }
-        ]
+              "  this.outstandingEvents_ = new goog.structs.Map();\n                                ^^^^^^^^^^^^^^^^",
+          },
+        ],
       });
       assert.equal(
         actual,
@@ -77,9 +77,9 @@ ${command}
             source:
               "/path/to/node_modules/google-closure-library/closure/goog/debug/tracer.js",
             line: 57,
-            column: 32
-          }
-        ]
+            column: 32,
+          },
+        ],
       });
       assert.equal(
         actual,